update tests and styling

+++ app.py
@@ -1,4 +1,3 @@
-import json
 from flask import Flask, render_template, redirect, flash, redirect, session, g, abort, request, jsonify
 from flask_debugtoolbar import DebugToolbarExtension
 import requests
@@ -10,12 +9,16 @@ import os
 
 app = Flask(__name__)
 
-app.config['SECRET_KEY'] = 'this is a secret'
 app.config['SQLALCHEMY_DATABASE_URI'] = (
     os.environ.get('DATABASE_URL', 'postgresql:///camphub'))
+app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY', 'camhub_key')
+
+# delete these when you're ready to push to Heroku
 toolbar = DebugToolbarExtension(app)
 app.config["SQLALCHEMY_ECHO"] = False
 app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+app.config["DEBUG_TB_INTERCEPT_REDIRECTS"] = False
+
 
 CURR_USER_KEY = "curr_user"
 base_url = "https://public-api.wordpress.com/rest/v1.1/sites"
@@ -59,7 +62,7 @@ def landing_page():
     '''Show landing page.'''
 
     if g.user:
-      return redirect(f"/home/{g.user.id}")
+        return redirect(f"/home/{g.user.id}")
 
     return render_template("landing_page.html")
 
@@ -68,36 +71,36 @@ def landing_page():
 def camphub_info():
     '''Render camphub breakdown.'''
 
-    return render_template("breakdown.html")
+    return render_template("user_routes/whatIs.html")
     
 
 @app.route("/signup", methods = ["GET", "POST"])
 def signup():
-  '''Generates signup form and creates new user upon post request.'''
+    '''Generates signup form and creates new user upon post request.'''
 
-  form = Signup_Form()
+    form = Signup_Form()
 
-  if form.validate_on_submit():
+    if form.validate_on_submit():
 
-    try:
-        username = form.username.data
-        password = form.password.data
-        school_name = form.school_name.data
-        field_of_study = form.field_of_study.data
-      
-        # hashes given password to securly store in db. Return new user instance w/ hashed data. Add to db.session.
-        user = User.register(username, password, school_name, field_of_study)
+        try:
+            username = form.username.data
+            password = form.password.data
+            school_name = form.school_name.data
+            field_of_study = form.field_of_study.data
+        
+            # hashes given password to securly store in db. Return new user instance w/ hashed data. Add to db.session.
+            user = User.register(username, password, school_name, field_of_study)
 
-        db.session.commit()
-        complete_login(user)
+            db.session.commit()
+            complete_login(user)
 
-        return redirect(f"/home/{user.id}")
+            return redirect(f"/home/{user.id}")
 
-    except IntegrityError:
-        flash("Username already exists- please choose a new username.")
-        return render_template("user_routes/signup.html", form = form)
-  
-  return render_template("user_routes/signup.html", form = form)
+        except IntegrityError:
+            flash("Username already exists- please choose a new username.")
+            return render_template("user_routes/signup.html", form = form)
+
+    return render_template("user_routes/signup.html", form = form)
 
 
 @app.route("/login", methods = ["GET", "POST"])
@@ -116,11 +119,11 @@ def login():
         given_password = form.password.data
 
         user = User.authenticate(username, given_password)
-  
+
         if user:
-                flash(f"Welcome {user.username}!")
+                flash(f"Welcome, {user.username}!")
                 complete_login(user)
-      
+
                 return redirect (f"/home/{user.id}")
 
         elif not user:
@@ -133,7 +136,7 @@ def login():
 def logout():
 
     if not g.user:
-      return redirect("/signup")
+        return redirect("/signup")
 
     del session[CURR_USER_KEY]
     return redirect("/camphub")
@@ -166,22 +169,25 @@ def edit_profile(user_id):
                 return redirect(f"/home/{user_id}")
 
         except:
-          flash ("Please check your spelling and try again.")
-          return redirect(f"/edit/profile/{user_id}")
+            flash ("Please check your spelling and try again.")
+            return redirect(f"/edit/profile/{user_id}")
 
     return render_template("user_routes/editProfile.html", form = form, user = user)
 
 
-@app.route("/delete/user/<int:user_id>")
+@app.route("/delete/user/<int:user_id>", methods = ["POST"])
 def delete_user(user_id):
     '''Delete user account.'''
 
     if not g.user:
         return redirect("/signup")
 
-    try:
-        user = User.query.get_or_404(user_id)
+    user = User.query.get_or_404(user_id)
 
+    if g.user.id != user.id:
+        return redirect("/camphub")
+
+    try:
         db.session.delete(user)
         db.session.commit()
 
@@ -204,7 +210,7 @@ def render_homepage(user_id):
 
     if user != g.user:
         return redirect(f"/home/{g.user.id}")
-  
+
     return render_template("user_routes/home.html")
 
 @app.route("/camphub/direction")
@@ -236,21 +242,20 @@ def view_user_posts():
     '''View all camphub user posts.'''
 
     if not g.user:
-      return redirect("/")
+        return redirect("/")
 
     all_posts = Camphub_User_Post.query.all()
-    print("******ALL POSTS ARE*******")
-    print(all_posts)
 
     return render_template("user_post_routes/all_posts.html", all_posts = all_posts)
+
 
 @app.route("/view/camphub/<int:post_id>")
 def view_given_post(post_id):
     '''Display given camphub post to authorized user.'''
 
     if not g.user:
-      flash("Unauothorzed access- please signup, or login if you have an account.")
-      return redirect("/signup")
+        flash("Unauothorzed access- please signup, or login if you have an account.")
+        return redirect("/signup")
 
     post = Camphub_User_Post.query.filter_by(id = post_id).first_or_404()
 
@@ -266,11 +271,7 @@ def create_user_post(user_id):
     if not g.user:
         return redirect("/")
 
-    user = User.query.get(user_id)
-
-    if g.user != user:
-        flash("Unauhorized route.")
-        return redirect(f"/home/{g.user.id}")
+    user = User.query.get_or_404(user_id)
 
     form = Camphub_User_Post_Form()
 
@@ -301,18 +302,14 @@ def create_user_post(user_id):
     return render_template("user_post_routes/create_post.html", form = form)
 
 
-@app.route("/camphub/delete/post/<int:post_id>")
+@app.route("/camphub/delete/post/<int:post_id>", methods = ["POST"])
 def delete_post(post_id):
     '''Delete the given user post and post comments.'''
 
     if not g.user:
         return redirect("/signup")
 
-    post = Camphub_User_Post.query.filter_by(id = post_id, author_id = g.user.id).first()
-
-    if not post:
-        flash("Post does not exist, so it cannot be deleted.")
-        return redirect("/camphub/users/posts")
+    post = Camphub_User_Post.query.filter_by(id = post_id, author_id = g.user.id).first_or_404()
 
     try: 
         db.session.delete(post)
@@ -334,7 +331,7 @@ def view_camphub_comments(user_id):
     '''View comments made here on camphub- does not include Wordpress Comments.'''
 
     if not g.user:
-      return redirect("/")
+        return redirect("/")
 
     user = User.query.get_or_404(user_id)
 
@@ -345,68 +342,48 @@ def view_camphub_comments(user_id):
     return render_template("user_post_routes/camphub_comments.html", all_comments = all_comments)
 
 
-# view one comment- should I have this 
-
-@app.route("/view/<int:post_id>/<int:comment_id>")
-def view_given_comment(post_id, comment_id):
-  return "this route needs work"
-
-
 @app.route("/create/comment/<int:post_id>/<int:user_id>", methods = ["GET", "POST"])
 def make_post_comment(post_id, user_id):
-      '''Allow authorized user to create a camphub (in-app) comment to an existing IN-APP post.'''
+    '''Allow authorized user to create a camphub (in-app) comment to an existing IN-APP post.'''
 
-      if not g.user:
-          return redirect("/")
+    if not g.user:
+        return redirect("/")
 
-      form = Camphub_Comment_Form()
+    form = Camphub_Comment_Form()
 
-      post = Camphub_User_Post.query.filter_by(id = post_id).first()
-      user = User.query.get(user_id)
+    post = Camphub_User_Post.query.get_or_404(post_id)
+    user = User.query.get_or_404(user_id)
 
-      if g.user != user:
-        flash("Unauthorized route.")
-        return redirect(f"/home/{g.user.id}")
+    if form.validate_on_submit():
+        try:
+            comment_user_id = user_id
+            camphub_post_id = post_id
+            content = form.content.data
 
-      if not post and user: 
-          flash("Post or user do not exist.")
-          return redirect("/camphub/users/posts")
+            new_post_comment = Camphub_Comment(comment_user_id = comment_user_id, camphub_post_id = camphub_post_id, content= content)
 
-      if form.validate_on_submit():
-          try:
-              print("form was submitted")
-              comment_user_id = user_id
-              camphub_post_id = post_id
-              content = form.content.data
+            db.session.add(new_post_comment)
+            db.session.commit()
 
-              new_post_comment = Camphub_Comment(comment_user_id = comment_user_id, camphub_post_id = camphub_post_id, content= content)
+            flash("Created Comment!")
 
-              db.session.add(new_post_comment)
-              db.session.commit()
-              
-              flash("Created Comment!")
+            return redirect(f"/view/camphub/{post_id}")
 
-              return redirect(f"/view/camphub/{post_id}")
+        except: 
+            flash("Something went wrong - please try again.")
+            return redirect(f"/view/camphub/{post_id}")
 
-          except: 
-              flash("Something went wrong - please try again.")
-              return redirect(f"/view/camphub/{post_id}")
-              
-      return render_template("/user_post_routes/new_comment.html", form = form, post = post)
+    return render_template("/user_post_routes/new_comment.html", form = form, post = post)
 
 
-@app.route("/camphub/delete/<int:post_id>/<int:comment_id>")
+@app.route("/camphub/delete/<int:post_id>/<int:comment_id>", methods = ["POST"])
 def delete_post_comment(post_id, comment_id):
     '''Delete given comment of given user post.'''
 
     if not g.user:
-      return redirect("/signup")
+        return redirect("/signup")
 
-    delete_comment = Camphub_Comment.query.filter_by(camphub_post_id = post_id, id = comment_id).first()
-
-    if not delete_comment:
-      flash("Unable to delete a post or comment that does not exist.")
-      return redirect("/camphub/users/posts")
+    delete_comment = Camphub_Comment.query.filter_by(camphub_post_id = post_id, id = comment_id).first_or_404()
 
     try:
         db.session.delete(delete_comment)
@@ -416,8 +393,8 @@ def delete_post_comment(post_id, comment_id):
         return redirect(f"/view/camphub/{post_id}")
 
     except:
-          flash("Something went wrong- please try agaib.")
-          return (f"/view/camphub/{post_id}")
+        flash("Something went wrong- please try agaib.")
+        return (f"/view/camphub/{post_id}")
 
 
 # # # # # # # # # # # # # # # # # # # # # # # # Suggest a Topic 
@@ -431,39 +408,26 @@ def suggest_topic(user_id):
 
     user = User.query.get_or_404(user_id)
 
+    if g.user != user:
+        return redirect("/")
+
     form = Suggest_Topic_Form()
 
     if form.validate_on_submit():
-      try :
-        print("try **************************ran")
+        try :
+            topic = form.topic.data
+            details = form.details.data
 
-        topic = form.topic.data
-        print("try **************************ran")
+            suggestion = Suggest_Topic(user_id = g.user.id, topic = topic, details = details)
 
-        print(topic)
+            db.session.add(suggestion)
+            db.session.commit()
 
-        details = form.details.data
-        print("try **************************ran")
+            flash("Suggestion submitted!")
+            return redirect(f"/home/{user_id}")
+        except:
+            flash("Something went wrong- please try again.")
 
-        print(details)
-
-        new_suggestion = Suggest_Topic(user_id = user_id, topic = topic, details = details)
-
-        print(type(Suggest_Topic))
-        print(Suggest_Topic)
-
-        print("new_suggestion is: *********")
-        print(new_suggestion)
-
-        db.session.add(new_suggestion)
-        db.session.commit()
-
-        flash("Request was submitted!")
-        return redirect(f"/home/{user_id}")
-      except:
-          flash("Something went wrong- please try again.")
-
-      
     return render_template("user_routes/suggestTopic.html", form = form)
 
 
@@ -480,20 +444,18 @@ def camphub_posts():
     if not g.user:
         redirect("/")
 
-    print("********* posts is ****************")
-
     resp = requests.get(f"{base_url}/{camphub_site}/posts/")
     resp = resp.json()
     
     articles = resp["posts"]
+    content = {}
     
     #  This needs work - the tags are removed w/ loop but unable to render on pg- something to do w/ the endfor?
 
     for i in range(len(resp["posts"])):
-      final_content = (resp["posts"][i]["content"][4:-5])
-      print(final_content)
-
-
+        final_content = (resp["posts"][i]["content"][4:-5])
+        print(final_content)
+        
     print("CONTENT IS: ********")
 
     articles = sorted(articles, key=lambda d: d['date'])
@@ -514,21 +476,11 @@ def view_article_CH_comment(article_id):
         resp = requests.get(f"{base_url}/{camphub_site}/posts/{article_id}")
         article = resp.json()
 
-        print("article ran")
-        # print(article)
-
         replies = requests.get(f"{base_url}/{camphub_site}/posts/{article_id}/replies/")
         replies = replies.json()
-        print("replies ran")
         replies = replies["comments"]
 
-        print("*************************")
-        # print(replies)
-
         user_comments_on_WP_article = Wordpress_Post_Comment.query.filter_by(wordpress_article_id = article_id).all()
-
-        print("this is in house")
-        print(user_comments_on_WP_article)
 
         return render_template("wp_in_app_routes/single_article.html", article = article, replies = replies, user_comments_on_WP_article = user_comments_on_WP_article)
 
@@ -541,7 +493,7 @@ def create_WP_camphub_comment(article_id):
     '''Create new camphub comment - EXCLUSIVELY ON MODERATOR POSTS.'''
 
     if not g.user:
-      redirect("/")
+        redirect("/")
 
     try:
         resp = requests.get(f"{base_url}/{camphub_site}/posts/{article_id}")
@@ -551,33 +503,32 @@ def create_WP_camphub_comment(article_id):
     except:
         flash("Article does not exist.")
         return redirect("/wordpress/articles/all")
-      
+
     form = Wordpress_CH_Article_Comment_Form()
 
     if form.validate_on_submit():
 
         try: 
-          wordpress_article_id = article_id
-          user_id = g.user.id
-          user_comment = form.user_comment.data
+            wordpress_article_id = article_id
+            user_id = g.user.id
+            user_comment = form.user_comment.data
 
-          new_comment = Wordpress_Post_Comment(wordpress_article_id = wordpress_article_id, user_id = user_id, user_comment = user_comment)
+            new_comment = Wordpress_Post_Comment(wordpress_article_id = wordpress_article_id, user_id = user_id, user_comment = user_comment)
 
-          db.session.add(new_comment)
-          print(new_comment)
-          db.session.commit()
+            db.session.add(new_comment)
+            db.session.commit()
 
-          flash("Your comment was added to CampHub comments.")
-          return redirect(f"/wordpress/camphub/article/{article_id}")
+            flash("Your comment was added to CampHub comments.")
+            return redirect(f"/wordpress/camphub/article/{article_id}")
 
         except:
-          flash("Something went wrong- please try again.")
-          return redirect(f"/create/comment/{article_id}")
+            flash("Something went wrong- please try again.")
+            return redirect(f"/create/comment/{article_id}")
 
 
     return render_template("wp_in_app_routes/new_comment.html", form = form, article = article, article_id = article_id)
 
-@app.route("/wordpress/delete/<int:article_id>/<int:comment_id>")
+@app.route("/wordpress/delete/<int:article_id>/<int:comment_id>", methods = ["POST"])
 def delete_wordpress_comment(article_id, comment_id):
     '''Delete user in-app wordpress comment.'''
 
@@ -587,7 +538,7 @@ def delete_wordpress_comment(article_id, comment_id):
     comment_to_delete = Wordpress_Post_Comment.query.filter_by(id = comment_id, wordpress_article_id = article_id).first()
 
     if not comment_to_delete:
-      return redirect("/wordpress/articles/all")
+        return redirect("/wordpress/articles/all")
 
     try:
         db.session.delete(comment_to_delete)
